btree/tests: Move persist test to a dedicated file
